src/email_channel/pyproject.toml
[tool.poetry]
name = "email-client"
version = "0.1.0"
description = ""
authors = ["Yaroslav PIsarev <yaricp@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.4"
loguru = "^0.7.2"
six = "^1.17.0"
kafka-python = "^2.0.2"
protobuf = "^5.29.1"
pydantic = "2.9.2"
pydantic-extra-types = "2.10.0"
pydantic-settings = "2.6.1"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
